+++ CHANGELOG.rst
@@ -35,6 +35,11 @@ Changed
 
 - Updated parser version to 10.
 
+- Replace backend to implement DFT-D3 dispersion correction.
+  Use `s-dftd3 <https://github.com/awvwgk/simple-dftd3>`_ instead of
+  `dftd3-lib <https://github.com/dftbplus/dftd3-lib>`_.
+  Option ``WITH_DFTD3`` is removed and replaced with ``WITH_SDFTD3``.
+
 Fixed
 -----